chore(services): split pipeline/candidate service files under 300-line cap
Closes #20. Splits by subdomain, matching each domain's existing router
file boundaries (pool.ts/tags.ts/documents.ts, stages.ts/analytics.ts):

- pipeline.service.ts (354->180) -> + pipeline-stages.service.ts,
  pipeline-analytics.service.ts
- candidate.service.ts (391->279) -> + candidate-pool.service.ts,
  candidate-tags.service.ts, candidate-documents.service.ts
- candidate-assessment.service.ts (311->198) -> +
  candidate-assessment-lifecycle.service.ts (submitAssessment's single
  write transaction can't be split further, so it stays alone)

No behavior change — pure extraction, router call sites and tests updated
to match.
