+++ packages/api/src/routers/pipeline/analytics.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod';
 import { router, permissionProcedure } from '../../trpc';
-import { pipelineService } from '../../services/pipeline.service';
+import { pipelineAnalyticsService } from '../../services/pipeline-analytics.service';
 import { assertScoped } from '../../access';
 
 export const pipelineAnalyticsRouter = router({
@@ -9,7 +9,7 @@ export const pipelineAnalyticsRouter = router({
     .input(z.object({ vacancyId: z.string().uuid() }))
     .query(async ({ ctx, input }) => {
       await assertScoped('vacancy', input.vacancyId, ctx.access, ctx.user.id, ctx.user.organizationId);
-      return pipelineService.getSlaStatus(ctx.user.organizationId, input.vacancyId);
+      return pipelineAnalyticsService.getSlaStatus(ctx.user.organizationId, input.vacancyId);
     }),
 
   // getNextBestAction is per-application — probe the application first.
@@ -17,13 +17,13 @@ export const pipelineAnalyticsRouter = router({
     .input(z.object({ applicationId: z.string().uuid() }))
     .query(async ({ ctx, input }) => {
       await assertScoped('application', input.applicationId, ctx.access, ctx.user.id, ctx.user.organizationId);
-      return pipelineService.getNextBestAction(ctx.user.organizationId, input.applicationId);
+      return pipelineAnalyticsService.getNextBestAction(ctx.user.organizationId, input.applicationId);
     }),
 
   getFunnel: permissionProcedure('pipeline', 'read')
     .input(z.object({ vacancyId: z.string().uuid() }))
     .query(async ({ ctx, input }) => {
       await assertScoped('vacancy', input.vacancyId, ctx.access, ctx.user.id, ctx.user.organizationId);
-      return pipelineService.getFunnel(ctx.user.organizationId, input.vacancyId);
+      return pipelineAnalyticsService.getFunnel(ctx.user.organizationId, input.vacancyId);
     }),
 });
